LockService as an interface
